Refactor the logic to find the dir containing blib or lib so it's more clear
The regex was hard to parse initially and had odd requirements like a
single non-whitespace character before "blib" or "lib".  The whole point
is to find the containing directory of blib or lib, and there are more
intuitive ways to handle that portably by treating the path as an array.

Note: the regex found the parent of a directory containing the words
"blib" or "lib", meaning that "my-lib" would match.  This refactor
preserves that quirk by anchoring the regex /\bb?lib\b/ only at word
boundaries and not string start/end.